feat: add data validation in methods

+++ api/Controllers/CommentController.cs
@@ -21,13 +21,19 @@ namespace api.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(comment => comment.ToCommentDto());
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if(comment == null)
@@ -36,8 +42,11 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentRequestDto commentDto){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             if(!await _stockRepo.StockExists(stockId))
                 return BadRequest("Stock does not exist");
                 
@@ -47,8 +56,11 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if(comment == null)
@@ -57,8 +69,11 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var comment = await _commentRepo.DeleteAsync(id);
 
             if(comment == null)
