+++ api/Controllers/StockController.cs
@@ -21,13 +21,19 @@ namespace api.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if(stock == null)
@@ -38,15 +44,20 @@ namespace api.Controllers
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
+            if(!ModelState.IsValid)
+                return BadRequest();
             
+
             var stockModel = stockDto.ToStockFromCreateDto();
             await _stockRepo.CreateAsync(stockModel);
 
             return CreatedAtAction(nameof(GetById), new {id = stockModel.Id}, stockModel.ToStockDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto){
+            if(!ModelState.IsValid)
+                return BadRequest();
 
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
@@ -57,8 +68,11 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if(!ModelState.IsValid)
+                return BadRequest();
+
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if(stockModel == null)
